poster-3-object-detection/requirements.txt
torch
rich
torchvision
pillow
glob2
wandb
matplotlib
tensordict
jason
opencv-python